Player Melee Projectile Pool Implementation
Implementation currently uses some placeholders, like a placeholder sprite, very basic code, and generic names for the pool at the moment. There is currently no collision detection on the melee projectile at the moment, and will need to be implemented in a way that calls the TakeDamage() method on the Enemy script.